requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
email-validator==2.1.1
aiodns==3.1.1
aiosmtplib==2.0.2
SQLAlchemy==2.0.30
python-multipart==0.0.9